Refactor dataAdapter to file and add tests
