+++ packages/sophtron-adapter/src/createVc.test.ts
@@ -2,7 +2,7 @@ import { VCDataTypes } from "@repo/utils";
 
 import type { AdapterDependencies } from "./models";
 import { logClient } from "./test/utils/logClient";
-import { aggregatorCredentials } from "./test/testData/credentials";
+import { aggregatorCredentials } from "test/testData/aggregatorCredentials";
 import { createSophtronVC } from "./createVc";
 
 import {
