+++ packages/sophtron-adapter/src/test/testData/aggregatorCredentials.ts
